adds antagonists to the game (#18514)
* commit 1 - get me out

* she lives

* adds wizards

* thing

* surprise end hits take 1

* s

* d

* surprise end hits take 2

* montreal

* REAl

* strangelight

* guilford fall

* natural disasters

* envelope

* h

* lady elect

* test 321

* test 123


* hot toes test

* ss

* redundancy

* s

* test 2

* s²

* s³

* s²³

* a pray!

* life and limb

* epic problem

* hmm

* update

* fixes something

* fixes something²

* adds slaughter demons to the game

* rend it

* hmm

* restores something

* adds clockwork cult into the game

* adds changelings to the game

* cassevetes

* test 101

* :)

* against

* shut the door

* adds darkspawn to the game

* sad

* cashout

* adds vampires to the game

* 2

* summer freeze

* pink frosty

* test111

* adds game to the game

* 2

* syndrome

* test

* test 2

* test 3

* test 4

* adds replay to the game?

* maybe?

* slo

* hrn

* test II

* test III

* test IV

* new technique

* ahm hum

* d

* sensible

* c

* ss13

* a

* v

* f

---------

+++ code/datums/mutations.dm
@@ -12,7 +12,8 @@
 	var/text_gain_indication = ""
 	var/text_lose_indication = ""
 	var/static/list/list/mutable_appearance/visual_indicators = list()
-	var/obj/effect/proc_holder/spell/power
+	/// The path of action we grant to our user on mutation gain
+	var/datum/action/cooldown/spell/power_path
 	var/layer_used = MUTATIONS_LAYER //which mutation layer to use
 	var/list/species_allowed = list() //to restrict mutation to only certain species
 	var/health_req //minimum health required to acquire the mutation
@@ -81,9 +82,9 @@
 		owner.remove_overlay(layer_used)
 		owner.overlays_standing[layer_used] = mut_overlay
 		owner.apply_overlay(layer_used)
-	grant_spell() //we do checks here so nothing about hulk getting magic
+	grant_power() //we do checks here so nothing about hulk getting magic
 	if(!modified)
-		addtimer(CALLBACK(src, PROC_REF(modify), 5)) //gonna want children calling ..() to run first
+		addtimer(CALLBACK(src, PROC_REF(modify), 0.5 SECONDS)) //gonna want children calling ..() to run first
 
 /datum/mutation/human/proc/get_visual_indicator()
 	return
@@ -109,8 +110,10 @@
 			mut_overlay.Remove(get_visual_indicator())
 			owner.overlays_standing[layer_used] = mut_overlay
 			owner.apply_overlay(layer_used)
-		if(power)
-			owner.RemoveSpell(power)
+		if(power_path)
+			// Any powers we made are linked to our mutation datum,
+			// so deleting ourself will also delete it and remove it
+			// ...Why don't all mutations delete on loss? Not sure.
 			qdel(src)
 		return 0
 	return 1
@@ -136,12 +139,21 @@
 				overlays_standing[CM.layer_used] = mut_overlay
 				apply_overlay(CM.layer_used)
 
+/**
+ * Called when a chromosome is applied so we can properly update some stats
+ * without having to remove and reapply the mutation from someone
+ *
+ * Returns `null` if no modification was done, and
+ * returns an instance of a power if modification was complete
+ */
 /datum/mutation/human/proc/modify() //called when a genome is applied so we can properly update some stats without having to remove and reapply the mutation from someone
-	if(modified || !power || !owner)
+	if(modified || !power_path || !owner)
 		return
-	power.charge_max *= GET_MUTATION_ENERGY(src)
-	power.charge_counter *= GET_MUTATION_ENERGY(src)
-	modified = TRUE
+	var/datum/action/cooldown/spell/modified_power = locate(power_path) in owner.actions
+	if(!modified_power)
+		CRASH("Genetic mutation [type] called modify(), but could not find a action to modify!")
+	modified_power.cooldown_time *= GET_MUTATION_ENERGY(src) // Doesn't do anything for mutations with energy_coeff unset
+	return modified_power
 
 /datum/mutation/human/proc/copy_mutation(datum/mutation/human/HM)
 	if(!HM)
@@ -170,15 +182,20 @@
 	else
 		qdel(src)
 
-/datum/mutation/human/proc/grant_spell()
-	if(!ispath(power) || !owner)
+/datum/mutation/human/proc/grant_power()
+	if(!ispath(power_path) || !owner)
 		return FALSE
 
-	power = new power()
-	power.action_background_icon_state = "bg_tech_blue_on"
-	power.panel = "Genetic"
-	owner.AddSpell(power)
-	return TRUE
+	var/datum/action/cooldown/new_power = new power_path(src)
+	new_power.background_icon_state = "bg_tech_blue"
+	new_power.base_background_icon_state = new_power.background_icon_state
+	new_power.active_background_icon_state = "[new_power.base_background_icon_state]_active"
+	new_power.overlay_icon_state = "bg_tech_blue_border"
+	new_power.active_overlay_icon_state = null
+	new_power.panel = "Genetic"
+	new_power.Grant(owner)
+
+	return new_power
 
 // Runs through all the coefficients and uses this to determine which chromosomes the
 // mutation can take. Stores these as text strings in a list.
